Data structure tests: Fixed code to work with new JUnit library (4.5.0) by adding deboxing casts.

+++ Tests/p79068/datastruct/QueueTest.java
@@ -158,7 +158,7 @@ public abstract class QueueTest {
 				enqueued++;
 			} else {
 				if (!queue.isEmpty()) {
-					assertEquals(dequeued, queue.dequeue());
+					assertEquals(dequeued, (int)queue.dequeue());
 					dequeued++;
 				}
 			}
